USB HID mouse support + fix a real signed-char movement bug
Mouse (Phase 3 follow-up): src/PicoUsbMouse.{hpp,cpp} parses the standard
3-byte boot-protocol report (buttons + signed dx/dy, no report-ID prefix),
sharing PicoUsbKeyboard's TinyUSB host stack -- that class now owns the
shared tuh_hid_*_cb callbacks since TinyUSB's C ABI allows only one
definition of each per program. Unlike TOM6809's own MouseState (an
accumulated absolute cursor position for an LVGL UI), DOOM's ev_mouse wants
relative deltas per tic; take_delta() accumulates raw HID units between
polls and resets on read. F12/F11/F10 (i_input_usbhid.cpp) toggle the mouse
and adjust DOOM's own mouseSensitivity, intercepted before reaching DOOM's
event system so they don't also trigger the vanilla quit-prompt/gamma-cycle
bindings on the same physical keys. Stats line now prints every 10s instead
of 1s.

Found and fixed the actual cause of "can't move backward, insensitive to
sensitivity" along the way -- a real, live instance of the char-signedness
difference the Phase 2 type audit had only found dead macros for until now:
d_ticcmd.h's ticcmd_t.forwardmove/sidemove were plain `char` (signed by
default on x86 Linux GCC, unsigned by default on this ARM toolchain).
G_BuildTiccmd's `cmd->forwardmove += forward` wrote small negative values
into an effectively-unsigned field, wrapping them to ~256 and reading back
as a large *positive* move -- backward became impossible, replaced by an
always-near-max forward, regardless of the original delta's magnitude.
G_ReadDemoTiccmd, ~30 lines away, already explicitly casts (signed char)
reading these same fields from a demo file, confirming the original
intent. Also fixed the same D_PostEvent/D_ProcessEvents sequence-point UB
(x = (++x) & mask, modifying x twice with no sequence point) already known
dead in d_net.c but very much live here -- ruled out first, real fix was
the ticcmd_t field type.

Followed up with a full type-size audit and a scoped <stdint.h> migration:
doomtype.h now pulls in <stdint.h>; every struct that actually crosses a
boundary (ticcmd_t, event_t, doomdata_t/doomcom_t, maptexture_t/mappatch_t
-- the WAD-format struct behind the earlier `masked` bug) is pinned to
explicit-width types instead of int/short/char/long. Deliberately not
touching the rest of the untouched-since-1993 engine, where int/short sizes
are already proven identical between this port's two platforms -- see
docs/PLAN.md for the full audit and the reasoning against a wider sweep.

+++ doom/d_net.h
@@ -57,18 +57,21 @@ typedef enum
 
 
 //
-// Network packet data.
+// Network packet data. Would be the actual wire format in a multiplayer
+// game (see docs/PLAN.md's type-size audit -- currently unreachable code,
+// i_net_null.c's doomcom fixes numnodes=1, but pinned to explicit widths
+// now rather than if/when that ever changes).
 //
 typedef struct
 {
     // High bit is retransmit request.
-    unsigned		checksum;
+    uint32_t		checksum;
     // Only valid if NCMD_RETRANSMIT.
-    byte		retransmitfrom;
-    
-    byte		starttic;
-    byte		player;
-    byte		numtics;
+    uint8_t		retransmitfrom;
+
+    uint8_t		starttic;
+    uint8_t		player;
+    uint8_t		numtics;
     ticcmd_t		cmds[BACKUPTICS];
 
 } doomdata_t;
@@ -76,54 +79,62 @@ typedef struct
 
 
 
+// This is the struct src/i_net_null.c actually fills in for real (see its
+// I_InitNetwork, mirroring i_net.c's own single-player setup path) -- a
+// genuine cross-boundary contract for this port, even without real
+// networking, so pinned the same as doomdata_t above. `long id` was the
+// one field worth specifically flagging: id==4 bytes here (ILP32 either of
+// this port's platforms), but `long` is 8 bytes on plenty of real-world
+// LP64 targets -- exactly the kind of assumption that's silently fine until
+// it isn't.
 typedef struct
 {
     // Supposed to be DOOMCOM_ID?
-    long		id;
-    
+    int32_t		id;
+
     // DOOM executes an int to execute commands.
-    short		intnum;		
+    int16_t		intnum;
     // Communication between DOOM and the driver.
     // Is CMD_SEND or CMD_GET.
-    short		command;
+    int16_t		command;
     // Is dest for send, set by get (-1 = no packet).
-    short		remotenode;
-    
+    int16_t		remotenode;
+
     // Number of bytes in doomdata to be sent
-    short		datalength;
+    int16_t		datalength;
 
     // Info common to all nodes.
     // Console is allways node 0.
-    short		numnodes;
+    int16_t		numnodes;
     // Flag: 1 = no duplication, 2-5 = dup for slow nets.
-    short		ticdup;
+    int16_t		ticdup;
     // Flag: 1 = send a backup tic in every packet.
-    short		extratics;
+    int16_t		extratics;
     // Flag: 1 = deathmatch.
-    short		deathmatch;
+    int16_t		deathmatch;
     // Flag: -1 = new game, 0-5 = load savegame
-    short		savegame;
-    short		episode;	// 1-3
-    short		map;		// 1-9
-    short		skill;		// 1-5
+    int16_t		savegame;
+    int16_t		episode;	// 1-3
+    int16_t		map;		// 1-9
+    int16_t		skill;		// 1-5
 
     // Info specific to this node.
-    short		consoleplayer;
-    short		numplayers;
-    
+    int16_t		consoleplayer;
+    int16_t		numplayers;
+
     // These are related to the 3-display mode,
     //  in which two drones looking left and right
     //  were used to render two additional views
     //  on two additional computers.
     // Probably not operational anymore.
     // 1 = left, 0 = center, -1 = right
-    short		angleoffset;
+    int16_t		angleoffset;
     // 1 = drone
-    short		drone;		
+    int16_t		drone;
 
     // The packet data to be sent.
     doomdata_t		data;
-    
+
 } doomcom_t;
